feat: add MultiViewObject3DUtils for resource disposal and integrate into PixiMultiViewManager and MultiViewPixi classes
