+++ src/MultiViewPixiBillboard.ts
@@ -1,13 +1,8 @@
 import { Container } from "pixi.js";
-import {
-  CanvasTexture,
-  Sprite,
-  Texture,
-  SpriteMaterial,
-  MeshBasicMaterial,
-} from "three";
+import { CanvasTexture, Sprite, Texture, SpriteMaterial } from "three";
 import { PixiMultiViewManager } from "./PixiMultiViewManager.js";
 import { IRenderablePixiView } from "./RenderablePixiView";
+import { MultiViewObject3DUtils } from "./MultiViewObject3DUtils.js";
 
 export class MultiViewPixiBillboard
   extends Sprite
@@ -76,35 +71,6 @@ export class MultiViewPixiBillboard
     MultiViewObject3DUtils.disposeMaterials(this.material);
     MultiViewObject3DUtils.disposeStageContainer(this._container);
     MultiViewObject3DUtils.disposeCanvas(this._canvas);
-  }
-}
-
-type MaterialType = MeshBasicMaterial | SpriteMaterial;
-
-export class MultiViewObject3DUtils {
-  static disposeMaterials(materials: MaterialType | MaterialType[]): void {
-    if (!Array.isArray(materials)) {
-      materials = [materials];
-    }
-    for (const material of materials) {
-      if (material.map) {
-        material.map.dispose();
-      }
-      material.dispose();
-    }
-  }
-
-  static disposeStageContainer(container: Container): void {
-    if (container) {
-      container.removeFromParent();
-      container.removeChildren();
-      container.destroy({ children: true });
-    }
-  }
-
-  static disposeCanvas(canvas: HTMLCanvasElement): void {
-    if (canvas && canvas.parentNode) {
-      canvas.parentNode.removeChild(canvas);
-    }
+    this.removeFromParent();
   }
 }
